\name{austria.cities}
\alias{austria.cities}
\docType{data}
\title{ Coordinates of Austrian cities }
\description{
  A dataset with geographic coordinates (longitude, latitude) of  
  Austrian cities (Landeshauptstaedte).

}
\usage{
  data(austria.cities)
}
\format{
  A two-column data frame with geographic longitudes and latitudes in
  the first and second column respectively. The data frame carries the
  grid type attribute, necessary for implicit working of transformations.

}
\details{

}
\source{
  This is the example dataset provided on:
  http://www.geoland.at/gps_trans.htm
}
\references{
  
}
\examples{
data(austria.cities)
attributes(austria.cities)
}
\keyword{datasets}
